+++ include/pika_set.h
@@ -15,7 +15,7 @@
  */
 class SAddCmd : public Cmd {
  public:
-  SAddCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SAddCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   std::vector<std::string> current_key() const override {
     std::vector<std::string> res;
     res.push_back(key_);
@@ -37,7 +37,7 @@ class SAddCmd : public Cmd {
 
 class SPopCmd : public Cmd {
  public:
-  SPopCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SPopCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   std::vector<std::string> current_key() const override {
     std::vector<std::string> res;
     res.push_back(key_);
@@ -60,7 +60,7 @@ class SPopCmd : public Cmd {
 
 class SCardCmd : public Cmd {
  public:
-  SCardCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SCardCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   std::vector<std::string> current_key() const override {
     std::vector<std::string> res;
     res.push_back(key_);
@@ -82,7 +82,7 @@ class SCardCmd : public Cmd {
 
 class SMembersCmd : public Cmd {
  public:
-  SMembersCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SMembersCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   std::vector<std::string> current_key() const override {
     std::vector<std::string> res;
     res.push_back(key_);
@@ -104,7 +104,7 @@ class SMembersCmd : public Cmd {
 
 class SScanCmd : public Cmd {
  public:
-  SScanCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag), pattern_("*") {}
+  SScanCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag), pattern_("*") {}
   std::vector<std::string> current_key() const override {
     std::vector<std::string> res;
     res.push_back(key_);
@@ -128,7 +128,7 @@ class SScanCmd : public Cmd {
 
 class SRemCmd : public Cmd {
  public:
-  SRemCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SRemCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   std::vector<std::string> current_key() const override {
     std::vector<std::string> res;
     res.push_back(key_);
@@ -151,7 +151,7 @@ class SRemCmd : public Cmd {
 
 class SUnionCmd : public Cmd {
  public:
-  SUnionCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SUnionCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override {};
   void Merge() override {};
@@ -165,7 +165,7 @@ class SUnionCmd : public Cmd {
 
 class SetOperationCmd : public Cmd{
  public:
-  SetOperationCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {
+  SetOperationCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {
     sadd_cmd_ = std::make_shared<SAddCmd>(kCmdNameSAdd, -3, kCmdFlagsWrite | kCmdFlagsSingleSlot | kCmdFlagsSet);
     del_cmd_ = std::make_shared<DelCmd>(kCmdNameDel, -2, kCmdFlagsWrite | kCmdFlagsMultiSlot | kCmdFlagsKv);
   }
@@ -191,7 +191,7 @@ class SetOperationCmd : public Cmd{
 
 class SUnionstoreCmd : public SetOperationCmd {
  public:
-  SUnionstoreCmd(const std::string& name, int arity, uint16_t flag) : SetOperationCmd(name, arity, flag) {}
+  SUnionstoreCmd(const std::string& name, int arity, uint32_t flag) : SetOperationCmd(name, arity, flag) {}
   // current_key() is override in base class : SetOperationCmd
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void DoUpdateCache(std::shared_ptr<Slot> slot = nullptr) override;
@@ -207,7 +207,7 @@ class SUnionstoreCmd : public SetOperationCmd {
 
 class SInterCmd : public Cmd {
  public:
-  SInterCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SInterCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override {};
   void Merge() override {};
@@ -220,7 +220,7 @@ class SInterCmd : public Cmd {
 
 class SInterstoreCmd : public SetOperationCmd {
  public:
-  SInterstoreCmd(const std::string& name, int arity, uint16_t flag) : SetOperationCmd(name, arity, flag) {}
+  SInterstoreCmd(const std::string& name, int arity, uint32_t flag) : SetOperationCmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void DoUpdateCache(std::shared_ptr<Slot> slot = nullptr) override;
   void DoThroughDB(std::shared_ptr<Slot> slot = nullptr) override;
@@ -235,7 +235,7 @@ class SInterstoreCmd : public SetOperationCmd {
 
 class SIsmemberCmd : public Cmd {
  public:
-  SIsmemberCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SIsmemberCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   std::vector<std::string> current_key() const override {
     std::vector<std::string> res;
     res.push_back(key_);
@@ -258,7 +258,7 @@ class SIsmemberCmd : public Cmd {
 
 class SDiffCmd : public Cmd {
  public:
-  SDiffCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SDiffCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override {};
   void Merge() override {};
@@ -271,7 +271,7 @@ class SDiffCmd : public Cmd {
 
 class SDiffstoreCmd : public SetOperationCmd {
  public:
-  SDiffstoreCmd(const std::string& name, int arity, uint16_t flag) : SetOperationCmd(name, arity, flag) {}
+  SDiffstoreCmd(const std::string& name, int arity, uint32_t flag) : SetOperationCmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void DoUpdateCache(std::shared_ptr<Slot> slot = nullptr) override;
   void DoThroughDB(std::shared_ptr<Slot> slot = nullptr) override;
@@ -286,7 +286,7 @@ class SDiffstoreCmd : public SetOperationCmd {
 
 class SMoveCmd : public Cmd {
  public:
-  SMoveCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {
+  SMoveCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {
     srem_cmd_ = std::make_shared<SRemCmd>(kCmdNameSRem, -3, kCmdFlagsWrite | kCmdFlagsSingleSlot | kCmdFlagsSet);
     sadd_cmd_ = std::make_shared<SAddCmd>(kCmdNameSAdd, -3, kCmdFlagsWrite | kCmdFlagsSingleSlot | kCmdFlagsSet);
   }
@@ -321,7 +321,7 @@ class SMoveCmd : public Cmd {
 
 class SRandmemberCmd : public Cmd {
  public:
-  SRandmemberCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  SRandmemberCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   std::vector<std::string> current_key() const override {
     std::vector<std::string> res;
     res.push_back(key_);
