+++ include/pika_transaction.h
@@ -13,7 +13,7 @@
 
 class MultiCmd : public Cmd {
  public:
-  MultiCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  MultiCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   Cmd* Clone() override { return new MultiCmd(*this); }
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override {}
@@ -25,7 +25,7 @@ class MultiCmd : public Cmd {
 
 class ExecCmd : public Cmd {
  public:
-  ExecCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  ExecCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   Cmd* Clone() override { return new ExecCmd(*this); }
   void Execute() override;
@@ -59,7 +59,7 @@ class ExecCmd : public Cmd {
 
 class DiscardCmd : public Cmd {
  public:
-  DiscardCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  DiscardCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   Cmd* Clone() override { return new DiscardCmd(*this); }
   void Split(std::shared_ptr<Slot> slot, const HintKeys& hint_keys) override {}
@@ -70,7 +70,7 @@ class DiscardCmd : public Cmd {
 
 class WatchCmd : public Cmd {
  public:
-  WatchCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  WatchCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
 
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   void Execute() override;
@@ -86,7 +86,7 @@ class WatchCmd : public Cmd {
 
 class UnwatchCmd : public Cmd {
  public:
-  UnwatchCmd(const std::string& name, int arity, uint16_t flag) : Cmd(name, arity, flag) {}
+  UnwatchCmd(const std::string& name, int arity, uint32_t flag) : Cmd(name, arity, flag) {}
 
   void Do(std::shared_ptr<Slot> slot = nullptr) override;
   Cmd* Clone() override { return new UnwatchCmd(*this); }
